refactor name paste for common
