+++ app/src/main/java/com/rendersoncs/report/common/util/DetailState.kt
@@ -1,4 +1,4 @@
-package com.rendersoncs.report.infrastructure.util
+package com.rendersoncs.report.common.util
 
 import com.rendersoncs.report.model.Report
