+++ app/src/main/java/com/rendersoncs/report/common/util/InternetCheck.java
@@ -1,4 +1,4 @@
-package com.rendersoncs.report.infrastructure.util;
+package com.rendersoncs.report.common.util;
 
 import android.os.AsyncTask;
