+++ app/src/main/java/com/rendersoncs/report/common/util/ResumeState.kt
@@ -1,4 +1,4 @@
-package com.rendersoncs.report.infrastructure.util
+package com.rendersoncs.report.common.util
 
 import com.rendersoncs.report.model.ReportResumeItems
