+++ app/src/main/java/com/rendersoncs/report/view/login/RecoveryLoginActivity.kt
@@ -14,7 +14,7 @@ import com.google.android.gms.tasks.Task
 import com.google.firebase.auth.FirebaseAuth
 import com.rendersoncs.report.R
 import com.rendersoncs.report.databinding.FragmentRecoveryLoginBinding
-import com.rendersoncs.report.infrastructure.util.closeVirtualKeyBoard
+import com.rendersoncs.report.common.util.closeVirtualKeyBoard
 
 class RecoveryLoginActivity : AppCompatActivity(), TextView.OnEditorActionListener {
     private var firebaseAuth: FirebaseAuth? = null
