+++ app/src/main/java/com/rendersoncs/report/view/login/SignUpActivity.kt
@@ -24,8 +24,8 @@ import com.google.firebase.database.DatabaseReference
 import com.google.firebase.storage.FirebaseStorage
 import com.rendersoncs.report.R
 import com.rendersoncs.report.databinding.FragmentSignUpBinding
-import com.rendersoncs.report.infrastructure.constants.ReportConstants
-import com.rendersoncs.report.infrastructure.util.closeVirtualKeyBoard
+import com.rendersoncs.report.common.constants.ReportConstants
+import com.rendersoncs.report.common.util.closeVirtualKeyBoard
 import com.rendersoncs.report.view.login.util.LibraryClass
 import com.rendersoncs.report.view.login.util.User
 import java.io.IOException
