+++ app/src/main/java/com/rendersoncs/report/view/viewmodel/ReportViewModel.kt
@@ -18,9 +18,9 @@ import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.ValueEventListener
 import com.rendersoncs.report.R
-import com.rendersoncs.report.infrastructure.constants.ReportConstants
-import com.rendersoncs.report.infrastructure.pdf.PDFGenerator
-import com.rendersoncs.report.infrastructure.util.*
+import com.rendersoncs.report.common.constants.ReportConstants
+import com.rendersoncs.report.common.pdf.PDFGenerator
+import com.rendersoncs.report.common.util.*
 import com.rendersoncs.report.model.*
 import com.rendersoncs.report.repository.ReportRepository
 import com.rendersoncs.report.view.base.SingleLiveEvent
